+++ docs/changelog.md
@@ -2,6 +2,17 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format. 
 
+## [0.8.0] - 2020-06-25
+
+### Added
+- plugins functionality
+
+### Changed
+- dropped Python 2 support
+
+### Removed
+- preciously deprecated `get_asset` method. Use `seek` instead
+
 ## [0.7.0] - 2020-03-17
 
 ### Added